requirements.txt
annotated-types==0.7.0
anyio==3.7.1
APScheduler==3.10.4
beautifulsoup4==4.12.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
coverage==7.11.0
Deprecated==1.3.1
dnspython==2.8.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.25.2
idna==3.11
iniconfig==2.3.0
limits==5.6.0
loguru==0.7.2
lxml==4.9.3
motor==3.3.2
packaging==25.0
pluggy==1.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pymongo==4.6.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
starlette==0.27.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==2.0.0
